HailDataFrame

The HailDataFrame class extends DataFrame to make a Hail Table object accessible through a data.frame-like interface.

Details

A HailDataFrame should more or less behave analogously to a data frame. It provides the same basic accessors (nrow, ncol, length, rownames, colnames, [, [<-, [[, [[<-, $, $<-, head, tail, etc) and can be coerced to an actual data frame via as.data.frame. Supported types of data manipulations include subset, transform, sort, xtabs, aggregate, unique, summary, etc.

A special feature of HailDataFrame, vs. an ordinary data frame, is that it can be grouped into a GroupedHailDataFrame, where every column is modeled as a list, split by some combination of grouping factors. This is useful for aggregation and supports the implementation of the aggregate method, which is the recommended high-level interface.

Another interesting feature is laziness: all column retrieval, e.g., via $ or eval, returns a HailPromise object. Many operations on promises are deferred, until they are finally fulfilled by being shown or through explicit coercion to an R vector.

Accessors

These are some accessors that HailDataFrame adds on top of the basic data frame accessors. Most of these are for advanced use only.

Extended API

Most of the typical data frame accessors and data manipulation functions will work analogously on HailDataFrame (see Details). Below, we list some of the non-standard methods that might be seen as an extension of the data frame API.

aggregate(x, data, FUN, ..., subset, na.action, simplify = TRUE, count = FALSE): If x is a formula, aggregates data, grouping by x, by either applying FUN, or evaluating an aggregating expression in …, on each group. If count is TRUE, a “count” column is added with the number of elements in each group. The rest of the arguments behave like those for the base aggregate. There are two main modes: aggregating with FUN, or, as an extension to the base aggregate, aggregating with expressions in …, similar to the interface for transform. If FUN is specified, then behavior is much like the original, except one can omit the LHS on the formula, in which case the entire frame is passed to FUN. In the second mode, there is a column in the result for each argument in …, and there must not be an LHS on the formula. See the documentation for the underlying facet function for details on what is supported on the formula RHS. For global aggregation, simply pass the HailDataFrame as x, in which case the data argument does not exist. Note that the function or expressions are only conceptually evaluated on each group. In reality, the computations occur on grouped columns/promises, which are modeled as lists. Thus, there is potential for conflict, in particular with length, which return the number of groups, instead of operating group-wise. One should use the abstraction ndoc instead of length, since ndoc always returns document counts, and thus will return the size of each group.

rename(x, ...): Renames the columns of x, where the names and character values of … indicates the mapping (newname = oldname).

group(x, by): Returns a GroupedHailDataFrame that is grouped by the factors in by, typically a formula. To get back to x, call ungroup(x).

grouping(x): Just returns NULL, since a HailDataFrame is not grouped (unless extended to be groupable).

defer(x): Returns a HailDataFrame that yields HailPromise objects instead of vectors whenever a field is retrieved

Getting data into Hail

There are two primary ways of getting a table of data into Hail:

Copy the data to Hail from an R data.frame-like object,

Instruct Hail to read the data from disk

See the example section below for a demonstration of both methods.

Evaluation

eval(expr, envir, enclos): Evaluates expr in the HailDataFrame envir, using enclos as the enclosing environment. The expr can be an R language object or a HailExpression, either of which are lazily evaluated if defer has been called on envir.

Coercion

as.data.frame(x, row.names=NULL, optional=FALSE): Downloads the data into an actual data.frame.

as.list(x): Essentially as.list(as.data.frame(x)), except returns a list of promises.

See also

Examples

library(hailr)
data_dir <- system.file("extdata", package="hailr")
tsv1 <- file.path(data_dir, "kt_example1.tsv")

## read directly from file into Hail
df <- readHailDataFrameFromText(tsv1, header=TRUE)

#> Error in hail_dir(): The rappdirs package must be installed to use the automaticallyinstalled Hail. Please install rappdirs or use a system Hail.

## read file into R, send to Hail
df <- send(read.table(tsv1, header=TRUE), hail())

#> Error in hail_dir(): The rappdirs package must be installed to use the automaticallyinstalled Hail. Please install rappdirs or use a system Hail.

## add/modify columns
transform(df, A = X / 2, B = X + 21)

#> Error in as.data.frame.default(x[[i]], optional = TRUE): cannot coerce class ‘"function"’ to a data.frame

### TOUGH: not supported by Hail directly, but could aggregate() and join()

transform(group(df, ~ SEX), HT_z = (HT - mean(HT)) / sd(HT))

#> Error in group(df, ~SEX): could not find function "group"

### The mean() and sd() calls would delegate to aggregate(), which
### produces an aggregated table. If we knew that the aggregated table
### was derived from the grouped table, then we could do the join as
### part of reconciling the contexts. It comes down to whether we
### should allow this type of syntax, with an implicit join from the
### aggregated to unaggregated data:

gdf <- group(df, ~ SEX)

#> Error in group(df, ~SEX): could not find function "group"

gdf$HT - mean(gdf$HT) / sd(gdf$HT)

#> Error in eval(expr, envir, enclos): object 'gdf' not found

### This is how a SplitDataFrameList would behave, but it is a list,
### so the behavior is sort of obvious. The implicit join may not be.

## filter rows
tsv2 <- file.path(data_dir, "kt_example2.tsv")
df2 <- readHailDataFrameFromText(tsv2, header=TRUE)

#> Error in hail_dir(): The rappdirs package must be installed to use the automaticallyinstalled Hail. Please install rappdirs or use a system Hail.

subset(df2, B != "rabbit")

#> Error in subset(df2, B != "rabbit"): object 'df2' not found

## aggregate
aggregate(df, mean_c1 = mean(C1), mean_c2 = mean(C2))

#> Error in as.data.frame.default(x): cannot coerce class ‘"function"’ to a data.frame

aggregate(~ SEX, df, mean_height_data = mean(HT))

#> Error in aggregate.formula(~SEX, df, mean_height_data = mean(HT)): 'formula' must have both left and right hand sides

## modify global fields
transform_globals(global_field_1 = c(1, 2, 3))

#> Error in transform_globals(global_field_1 = c(1, 2, 3)): could not find function "transform_globals"